Replaced Node server with updated Login server
- Added bodyparser to package.json (for parsing body of POST requests)
