+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "base-64": "^0.1.0",
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.17.2",
     "express": "^4.14.0",
     "jasmine": "^2.5.3",
     "jasmine-node": "^1.14.5",
