+++ server/index.js
@@ -3,7 +3,7 @@
 
 const app = require('./app');
 
-const PORT = process.env.PORT || 3000;
+const PORT = process.env.PORT || 3001;
 
 app.listen(PORT, () => {
   console.log(`App listening on port ${PORT}!`);
